requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.4.26
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.1
openai==1.82.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
pyzbar==0.1.9
sniffio==1.3.1
starlette==0.44.0
tqdm==4.67.1
typing-extensions==4.13.2
uvicorn==0.33.0
